services/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
redis==5.0.8
mysql-connector-python==9.0.0
prometheus-client==0.20.0
pydantic==2.9.2
pydantic[email]==2.9.2
